refactor component tests and expand coverage
